backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.2
cryptography==42.0.1
cssselect2==0.7.0
cycler==0.12.1
distro==1.9.0
dnspython==2.3.0
fastapi==0.115.12
firebase-admin==6.8.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
fonttools==4.57.0
google==3.0.0
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
grpcio-tools==1.71.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.0
kiwisolver==1.4.8
lxml==5.1.0
MarkupSafe==3.0.2
matplotlib==3.10.1
msgpack==1.1.0
numpy==2.2.5
openai==1.82.0
oscrypto==1.3.0
packaging==25.0
pandas==2.2.3
pdfkit==1.0.0
pillow==10.2.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.21
pydantic==2.11.4
pydantic_core==2.33.2
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyJWT==2.10.1
pymongo==4.5.0
pyparsing==3.2.3
pypdf==4.0.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.1
qrcode==7.4.2
redis==5.2.1
reportlab==4.0.9
requests==2.32.3
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.3
setuptools==78.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.46.2
svglib==1.5.1
threadpoolctl==3.6.0
tinycss2==1.2.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.1.0
uvicorn==0.34.2
webencodings==0.5.1
Werkzeug==3.1.3
xhtml2pdf==0.2.14
